+++ nowcasting_dataset/manager/__init__.py
@@ -1,0 +1,1 @@
+""" Data manager functions """
